Errores indeterminados de la App

+++ src/main/java/es/sd/practica1/RepositorioModelos.java
@@ -6,13 +6,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface RepositorioModelos extends JpaRepository<Modelos, Integer> {
 	
 	Modelos findByNombre (String nombre);
-	Modelos findByValorFacial (float valorFacial);
+	Modelos findByValorFacial (double valorFacial);
 	Modelos findByUnidad (String unidad);
 	Modelos findByDiametro (int diametro);
-	Modelos findByPeso (float peso);
+	Modelos findByPeso (double peso);
 	Modelos findByMetal (String metal);
 	Modelos findByDescripcion (String descripcion);
 	
-	List<Modelos> findByValorFacialAndUnidad (float valorFacial, String unidad); 
-
+	List<Modelos> findByValorFacialAndUnidad (double valorFacial, String unidad);
+	
 }
